+++ busy_beaver/tasks/retweeter.py
@@ -2,7 +2,7 @@ from datetime import timedelta
 import logging
 
 from busy_beaver import kv_store, slack, twitter
-from .toolbox import utc_now_minus
+from busy_beaver.toolbox import utc_now_minus
 
 LAST_TWEET_KEY = "last_posted_tweet_id"
 logger = logging.getLogger(__name__)
